// src/app/interfaces/client.ts
export interface Client{
    id: number;
    clientId: string;
    firstName: string;
    middleName: string;
    lastName: string;
    dateOfBirth: Date;
    assetValue: number;
    homePhoneNumber: string;
    officePhoneNumber: string;
    email: string;
    driversLicenseIdNum: string;
    userId: number;
    accountStatusId: number;
}